Split ROS TypeDB interface tests
